chore: add authentic source claim from EC
