+++ ts/features/itwallet/presentation/components/ItwPresentationClaimsSection.tsx
@@ -9,7 +9,7 @@ import { View } from "react-native";
 import I18n from "../../../../i18n";
 import { ItwCredentialClaim } from "../../common/components/ItwCredentialClaim";
 import { ItwQrCodeClaimImage } from "../../common/components/ItwQrCodeClaimImage";
-import { ItwReleaserName } from "../../common/components/ItwReleaserName";
+import { ItwIssuanceMetadata } from "../../common/components/ItwIssuanceMetadata";
 import {
   getCredentialStatus,
   parseClaims,
@@ -81,7 +81,7 @@ export const ItwPresentationClaimsSection = ({
         );
       })}
       {claims.length > 0 && <Divider />}
-      <ItwReleaserName credential={credential} isPreview={false} />
+      <ItwIssuanceMetadata credential={credential} isPreview={false} />
     </View>
   );
 };
